feat(api): extract ParcelCalculationService and refactor controller
- Isolates business logic: franchise thresholds, Chronopost tiered fees,
  EU country detection
- Keeps constants inline within the service (no external config)
- Controller reduced from ~150 lines to ~10 lines
